focus: make reading order `y` check use relative epsilon
For large magnitudes (e.g., stuff defined in `mm`) you'd want a higher
epsilon than small magnitudes (e.g., stuff defined in `km`). Using a
relative epsilon makes this somewhat more similar to a units-last-place
check. This also still isn't perfect, as when both are placed around
`0`, but one is slightly offset, this will treat the check as if we're
working in a very, very small magnitude.

In `kurbo`, for these types of things users need to pass in an explicit
tolerance, which in context of UIs can often be calculated as some
fraction of the pixel size. That's something we could do here, too.