box_tree: Change default clip behavior to `Inherit` (#68)
If a node has a clip, likely the most expected behavior is for no
children of that node to spill outside the clip (also if any children
have their own clips).

When `ClipBehavior` was introduced in
https://github.com/endoli/understory/pull/28, the default was set to the
old behavior of children with clips set to ignore the node's (their
parent's) clip.