index: Write benchmark against `Backend`, reuse and regroup (#85)
This suggests writing the usage patterns we're benchmarking against
`Backend`, and running them for all the backends. This ensures we
compare the same workload across backends.

This also suggests grouping the benchmark results under the usage
pattern first, and backend second, which allows comparing the backends
directly for each usage pattern.