box_tree: Walk the ancestor clips in `hit_test_point` (#95)
This checks all the local geometry to perform hit testing, checking the
node's local bounds, the local clip, and then the node's ancestors'
clips. Rounded rectangle clips are now handled correctly, rather than
checking point containment on the underlying `Rect`.

Because we need to walk the tree to correctly check the geometry, we do
away with the `depth` and `path_to_root` helpers (and build the path in
a scratch buffer as we're walking).

https://github.com/endoli/understory/pull/93 passes when based on this
PR.